+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 *Minor*
 
 * Removed the Devoted Servant. ([#127](https://github.com/hjwylde/werewolf/issues/127))
+* Renamed the Angel to the Fallen Angel. ([#130](https://github.com/hjwylde/werewolf/issues/130))
 
 *Revisions*
 
 * Changed the Druid's balance to 3. ([#165](https://github.com/hjwylde/werewolf/issues/165))
 * Changed the Wolf-hound's balance to -2. ([#167](https://github.com/hjwylde/werewolf/issues/167))
+* Updated the Fallen Angel's description and rules. ([#130](https://github.com/hjwylde/werewolf/issues/130))
 
 ### v0.5.2.0